chore(crate) reorganize structs
